Capstone Update

+++ app/Http/Controllers/Backend/Admin/SuppliersController.php
@@ -9,6 +9,7 @@ use App\{Supplier,User};
 use Auth;
 use Hash;
 use Image;
+use DB;
 
 class SuppliersController extends Controller
 {
@@ -47,6 +48,19 @@ class SuppliersController extends Controller
             'company'       => 'required|min:2|max:50',
             'email'         => 'nullable|email',
             'number'        => 'nullable|numeric|digits_between:0,11',
+        ],
+        [
+            'number.required'               => 'Contact Number is required',
+            'number.numeric'                => 'Contact Number should be numbers only.',
+            'number.digits_between'         => 'Contact Number should only contain 11 digits or lower',
+            'confirm_password.required'     => 'Confirm Password is required',
+            'confirm_password.same'         => 'Confirm Password and Password does not match',
+            'firstname.required'            => 'First Name field is required',
+            'firstname.min'                 => 'First Name should be atleast 2 characters',
+            'lastname.min'                  => 'Last Name should be atleast 2 characters',
+            'lastname.max'                  => 'Last Name maximum characters is 20',
+            'firstname.max'                 => 'First Name maximum characters is 20',
+            'lastname.required'             => 'Last Name field is required',
         ]);
 
         $supplier = new supplier();
@@ -207,26 +221,44 @@ class SuppliersController extends Controller
     {
         if(Hash::check(request('password'),Auth::user()->password)){
             $supplier = Supplier::find(request('supplier_id'));
-            $auth_user  = Auth::user();
-            $admins = User::where('role_id',1)->get();
-            $badge = array(
-                'bg' => "danger",
-                'icon' => "fa fa-address-card mx-0",
-            );
-            $supplier->delete();
-            
-            foreach($admins as $admin)
-            {
-                $admin->notify(new SupplierResource($auth_user,$supplier,"deleted",$badge));
-            }
+            $count = DB::table('products')
+                        ->where('supplier_id',$supplier->supplier_id)
+                        ->where('deleted_at','=',null)
+                        ->count();
 
-            $name = $supplier->getFullName();
-            $notification = array (
-                'message' => "Supplier ".$name." was successfuly deleted!",
-                'icon' => 'success',
-                'heading' => 'Success',
-            );
-            return back()->with($notification);
+            if($count == 0)
+            {
+                $auth_user  = Auth::user();
+                $admins = User::where('role_id',1)->get();
+                $badge = array(
+                    'bg' => "danger",
+                    'icon' => "fa fa-address-card mx-0",
+                );
+                $supplier->delete();
+                
+                foreach($admins as $admin)
+                {
+                    $admin->notify(new SupplierResource($auth_user,$supplier,"archived",$badge));
+                }
+
+                $name = $supplier->getFullName();
+                $notification = array (
+                    'message' => "Supplier ".$name." was successfuly deleted!",
+                    'icon' => 'success',
+                    'heading' => 'Success',
+                );
+                return back()->with($notification);
+            }
+            else
+            {
+                $notification = array(
+                    'message' => "Supplier is being used by other/s Product!",
+                    'icon'  => 'error',
+                    'heading'   => 'Failed!',
+                );
+                
+                return back()->with($notification);
+            }
         }
         else{
             $notification = array (
@@ -236,5 +268,40 @@ class SuppliersController extends Controller
             );
             return back()->with($notification);
         }
+    }
+
+    public function restore(Request $request)
+    {
+        $supplier = Supplier::onlyTrashed()
+                            ->where('supplier_id',request('supplier_id'))
+                            ->firstOrFail();
+        $auth_user = Auth::user();
+        $admins = User::where('role_id',1)->get();
+        $badge = array(
+            'bg' => "info",
+            'icon' => "fa fa-address-card mx-0",
+        );
+
+        $supplier->restore();
+
+        foreach($admins as $admin)
+        {
+            $admin->notify(new SupplierResource($auth_user,$supplier,"restored",$badge));
+        }
+       
+        $notification = array(
+            'message' => 'Supplier successfully Restored!',
+            'icon'    => 'success',
+            'heading' => 'SUccess!',
+        );
+
+        return back()->with($notification);
+    }
+
+    public function archive()
+    {
+        $suppliers = Supplier::onlyTrashed()->get();
+
+        return view('backend.admin.suppliers.suppliersArchive',compact('suppliers'));
     }
 }
